chat-bot-tarak-2/server/ai_core_service/requirements.txt
# Core Web Framework & Server
Flask
Flask-CORS
# waitress # Production WSGI server, consider using this or Gunicorn/Uvicorn

# Configuration & Utilities
python-dotenv
uuid # Standard library, listed for explicitness if previously included

# AI - Langchain Ecosystem
langchain
langchain-community
langchain-huggingface # For integrating HuggingFace models/embeddings with Langchain
langchain-ollama      # For Ollama integration with Langchain

# AI - LLM & Embeddings Libraries
ollama                # Direct Ollama client library
sentence-transformers # For generating sentence embeddings (primary for RAG)
google-generativeai
groq

# AI - Vector Store
# Choose one of the FAISS packages based on your environment:
faiss-cpu             # CPU version of FAISS
# faiss-gpu           # GPU version of FAISS (requires CUDA)

# Document Processing & Text Extraction
pypdf                 # Modern PDF library, successor to PyPDF2
PyPDF2                # Older PDF library, kept for compatibility if specific features were used
pymupdf               # PyMuPDF (Fitz) for efficient PDF parsing
python-docx           # For .docx (Word) files
python-pptx           # For .pptx (PowerPoint) files

# Tokenizers
tiktoken              # Tokenizer used by OpenAI models, often useful with Langchain

# Web Search & Content Scraping
duckduckgo-search
newspaper3k
lxml                  # For robust HTML parsing, required by newspaper3k
lxml_html_clean

# HTTP Client Libraries
requests              # General purpose HTTP client
httpx                 # Modern HTTP client, supports async

# Additional Dependencies for ai_core_service
redis

spacy                # For NLP tasks, tokenization, etc.

flask # Core Flask web framework
flask-cors # For handling CORS in Flask applications
neo4j # Neo4j database driver for Python
# Knowledge Graph & NLP
neo4j                 # Neo4j Python driver for knowledge graph
# After installing spacy via pip, you must download the language model separately.
# Run: python -m spacy download en_core_web_sm
spacy                 # Natural language processing library

# Additional Dependencies for ai_core_service
redis
# Additional Dependencies for ai_core_service
redis

python-dotenv
langchain
langchain-openai
langchain-groq

# requirements.txt
google-generativeai
langchain-google-genai
# pip install --upgrade langchain langchain-community

# pip install pymongo python-docx PyMuPDF python-pptx

# pip install transformers datasets peft accelerate bitsandbytes torch trl

# pip install langchain-groq

# pip install requests beautifulsoup4

# Requirements for Interactive Podcast Generator
pyttsx3
pydub
eyed3
PyPDF2
PyMuPDF
pytube
youtube-transcript-api
ollama
google-generativeai
groq

# Add gTTS to your server/ai_core_service/requirements.txt file.

# pip install python-pptx